Rename bookmarks_view package to bookmark
