+++ app/src/androidTestKiwix/java/org/kiwix/kiwixmobile/tests/activities/BookmarksActivityTest.java
@@ -9,7 +9,7 @@ import android.test.suitebuilder.annotation.LargeTest;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.kiwix.kiwixmobile.bookmarks_view.BookmarksActivity;
+import org.kiwix.kiwixmobile.bookmark.BookmarksActivity;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
