+++ app/src/main/AndroidManifest.xml
@@ -153,7 +153,7 @@
 
     <activity android:name=".settings.KiwixSettingsActivity" />
     <activity android:name=".search.SearchActivity" />
-    <activity android:name=".bookmarks_view.BookmarksActivity" />
+    <activity android:name=".bookmark.BookmarksActivity" />
 
     <provider
       android:name=".ZimContentProvider"
