+++ app/src/main/java/org/kiwix/kiwixmobile/bookmark/BookmarksPresenter.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.bookmarks_view;
+package org.kiwix.kiwixmobile.bookmark;
 
 import org.kiwix.kiwixmobile.ZimContentProvider;
 import org.kiwix.kiwixmobile.base.BasePresenter;
